+++ stack/app/rds.tf
@@ -23,8 +23,6 @@ module "rds" {
 
   project     = var.project
   environment = var.environment
-  account_id  = var.account_id
-  region      = var.region
   vpc_id      = module.vpc.id
   password    = var.rds_is_read_replica ? null : local.credentials.db_root_password
   kms_key_arn = var.rds_master_db_kms_key_arn == null ? module.rds-kms-key[0].arn : var.rds_master_db_kms_key_arn