+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project does not follow SemVer, since modules are independent of each other
 
 ## Unreleased
 
+### ecs-deploy/cluster
+* Add `regional` variables [#277](https://github.com/dbl-works/terraform/pull/277)
+
 ## [v2023.11.27]
 ### slack/ecr-scanner-notifier
 * Slack notifier for AWS Elastic Container Registry (ECR) image scans. It automatically sends notifications to a designated Slack channel whenever vulnerabilities are detected in ECR scans. [#196](https://github.com/dbl-works/terraform/pull/196)